refactor: Moved feature dropdown into FeatureControls
